requirements.txt
aioredis
feedparser
bs4
lxml
discord.py
ratelimiter
requests
aiohttp